Move constants for default values to const module
