+++ azely/consts.py
@@ -1,0 +1,8 @@
+# constants
+HERE = "here"
+NOW = "now"
+TODAY = "today"
+FRAME = "tcrs"
+FREQ = "10T"
+SEP = "to"
+TIMEOUT = 5
